cleanup build system
The Burp API files do not need to be present in the extension's JAR
because these interfaces already exist in Burp. These have been
removed from the source tree, compiled, and added to their own JAR
for use during development and building code. This JAR is not added
to the extension's JAR. Now, the only code you see in you IDE is the
code for this project.