The GUI has arrived!
I cleaned up the Database and Config classes and added extra
functionality to them to support GUI operations. The GUI code was
added, and I believe it is fully functional, with one exception:
saving the configuration. We have had code in place to save configs
for a long time, but I just have not called that code yet.