add issues based on observed json values
I pull out the associated keys as well, but I'm not doing anything with
them yet. Issues are created for the values. Number values are not yet
supported; that will come next. It's strings only for now.

Sample body that I've been testing with:
  "bp": {
        "_pl": 1,
        "A_v": "3.26.1",
        "_bt": "rapid",
        "A_pr": "https"
    },
To test, set the request content-type to application/json (or anything
containing json), and place the above in the body of the request.

This has not been fully tested because of a NullPointerException in
HashEngine.Hash which I have not tracked down yet. What I did test
seemed to work well, though. The JSON parsing is definitely working.